Move interpolation logic
